Puts activity and time processing of entries in its own file
